Always wait for the server greeting to arrive before sending anything.
We used to have a race condition built in there.
This additionally cleans up the logic a bit and contains it better in LoginJob.

Further we now always retrieve the CAPABILITIES to check the available
authentication methods, even if we're not on an encrypted connection
(this used to have an extra codepath).