Non-blocking stream parser.
Instead of blocking we wait for at least one CRLF to parse,
and if we fail to get to the end we simply abort and retry once we have
more data.

This way we no longer have to block the event loop, at the cost of
potentially parsing some bits multiple times. I suppose this could be
improved by using the literal size for better checks if all data is
available.