Reduce copying.
Instead of always trying to read as much as possible from the socket
into the buffer, and then resetting the buffer with the (potentially
large) remainder (which means we repeatedly memcpy a large portion),
we only read in small chunks when we need more data.
This way the remainder copy operation only has to copy a small chunk.
Further we don't reallocate the buffer all the time, and use two
buffers instead. This allows us to memcpy the remainder into the other
buffer, from which we can then read more data.

This reduced the memcpy load from ~40% to ~2%, although the benchmark
time only improved slightly.