API to avoid superfluous mime message parsing
In sink all we require is the encodedContent() binary blob which we're
dumping directly to disk and parse later. Therefore the parsing in kimap
is entirely superfluous and results in a major performance bottleneck.

While a more proper solution would be to just expose the binary content
and have the client parse it however it wants, this provides a quick fix
that will do for the time being.