Guard against recursive processing
When we deliver a message the callee has the opportunity to call the
parseStream() function recursively, which will make for a very confusing
internal state, since the internal logic is not prepared to handle that
case.

Right now we seem to rely on that case in Session, if QCoreApplication::processEvents
gets called via job-results from parseStream().
I'm not sure how because QIODevice::readyRead should not fire
recursively on the socket, but that's as far as I got.

Should we be able to rule out legitimate use of this guard then it
should be turned into an error.