Idempotent CREATE and DELETE
If CREATE fails with an ALREADYEXISTS response code, that is not an error
since the intended result (the object existing at the target location)
has been achieved.

Likewise, if DELETE fails with an NONEXISTENT response code, that is
not an error since the intended result (the object being gone from
the target location) has been achieved.

Please note that for RENAME response code ALREADYEXISTS *is* an error
since the content of the source folder was not moved as intended.

cf. https://tools.ietf.org/html/rfc5530#section-3

BUG: 423672