Enable keepalive on the socket
The theory is that this helps detecting if a connection is still
available (because of the keepalive packets). We don't actually care
about keeping the connection open for an extended time, but getting some
feedback on wether the connection is still alive would be great to avoid
first having to wait for a timeout before establishing a new connection.